requirements.txt
asgiref==3.8.1
bleach==6.2.0
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.4
djangorestframework==3.15.2
idna==3.10
pillow==11.1.0
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
webencodings==0.5.1
